MGMT-6812 Add parameters to track iso streaming to swagger
This adds `StreamImage` to the image create payload which effectively
turns the endpoint into just an update for the image info (no image is
created or uploaded).

Also, we store that value in the cluster in the new `IsStream` ImageInfo
property which will be used on download to determine that we should
create the customized iso during the stream rather that downloading it
from storage (it won't exist).