Allow other extra kernel params in preprovisioning images
Before this change the preprovisioningimage controller always set the
ExtraKernelParams field to only the rootfs parameter or empty depending
on the allowed image types.

This change accounts for user alterations to the field and prevents the
controller from overwriting them. This is specifically useful when
debugging where we may need a kernel breakpoint to assess network
conditions when the rootfs pull is failing.

Related to: https://issues.redhat.com/browse/OCPBUGS-1844