Ensure BMH is attached when the allow-provisioned-host-management annotation is set
Setting this annotation will keep the BMH attached as long as the converged
flow is enabled and the remove-agent-and-node-on-delete annotation is
not set.

If both annotations are set the remove-agent-and-node-on-delete takes
priority and sets detached as it's easier for a user to realize the host
is detached and deal with that than to find out the node isn't going to
be properly removed after the BMH has already started deprovisioning on
delete.

Resolves https://issues.redhat.com/browse/ACM-21001